+++ scripts/deploy.sh
@@ -11,11 +11,6 @@ elif [ "$TRAVIS_TAG" ]; then
 
   sh scripts/robottxt-gen.sh allow
   npm run deploy -- --env PROD;
-
-  # Easiest way to get embeds built and deployed only on a tagged commit.
-  rm -rf dist
-  npm run build:embeds
-  npm run deploy:embeds
 else
   echo ""
   echo "=> Deploying $TRAVIS_BRANCH to preview.gw2armory.com"